Fix compilation error
An earlier change modified the parameter type for messageSize to size_t,
but that caused a log statement to have the wrong specifier for
messageSize. Instead of changing the log statement, modify
sendMessageToNanoappFromHost to cast messageSize to a uint32_t before
sending to craftNanoappMessageFromHost since the latter method needs it
as a uint32_t anyway.

Test: ./load_chre.sh && ./run_sim.sh

+++ core/include/chre/core/host_comms_manager.h
@@ -198,7 +198,7 @@ class HostCommsManager : public NonCopyable {
                                                uint16_t hostEndpoint,
                                                uint32_t messageType,
                                                const void *messageData,
-                                               size_t messageSize);
+                                               uint32_t messageSize);
 
   /**
    * Posts a crafted event, craftedMessage, to a nanoapp for processing, and
